chore: move wallet ops to kadclient
- We didn't want a separate binary for this.
